requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
requests==2.31.0
Jinja2==3.1.3
MarkupSafe==2.1.3
Werkzeug==3.0.1
itsdangerous==2.1.2
gunicorn==21.2.0
apscheduler==3.10.4
